add: elevato-simulator files. fix: import errors
